+++ tilisp.h
@@ -51,9 +51,9 @@ struct value_t {
 
 value_t *display(const value_t *);
 
-void check_type(const value_t *, uint8_t);
-
+void check_type(const value_t *, uint64_t);
 void check_func(const value_t *, uint8_t);
+
 value_t *cpp_add(const value_t *value_1, const value_t *value_2);
 value_t *cpp_mult(const value_t *value_1, const value_t *value_2);
 value_t *cpp_div(const value_t *value_1, const value_t *value_2);
@@ -64,7 +64,6 @@ value_t *cpp_less_than(const value_t *value_1, const value_t *value_2);
 value_t *cpp_more_than(const value_t *value_1, const value_t *value_2);
 value_t *cpp_leq(const value_t *value_1, const value_t *value_2);
 value_t *cpp_geq(const value_t *value_1, const value_t *value_2);
-
 value_t *cpp_concat(const value_t *value_1, const value_t *value_2);
 }
